Moved InspectorTreeControllerBase and InspectorTreeFixedRowHeightController into inspector_tree_flutter.dart
